backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pandas==2.2.2
python-pptx==1.0.2
jinja2==3.1.4
tenacity==8.5.0
openai==1.43.0
